Refactor acl.js to prep for external solid-permissions lib
- Move `lib/acl.js`/`allow()` handler to `lib/handlers/allow`
- Extract ACL class to `lib/acl-checker.js` (convert to es6)
- Remove S() usage (use native String functions instead)
- Pass in the debug() logger as a dependency injection for ACLChecker
- Use new rdflib graph.match() natively
