+++ lib/handlers/get.js
@@ -10,7 +10,7 @@ var Negotiator = require('negotiator')
 
 var debug = require('debug')('solid:get')
 var debugGlob = require('debug')('solid:glob')
-var acl = require('../acl.js')
+var acl = require('./allow')
 
 var utils = require('../utils.js')
 var translate = require('../utils.js').translate
