+++ lib/ldp-middleware.js
@@ -2,7 +2,7 @@ module.exports = LdpMiddleware
 
 var express = require('express')
 var header = require('./header')
-var acl = require('./acl')
+var acl = require('./handlers/allow')
 var authentication = require('./handlers/authentication')
 var get = require('./handlers/get')
 var post = require('./handlers/post')
